backend/requirements.txt
Django==5.2.7
djangorestframework==3.15.2
mangum==0.17.0
psycopg2-binary==2.9.9